README.md
My website is a Windows 10 emulator created using HTML, CSS, and JavaScript. It replicates the Windows 10 user interface and includes some functional features that mimic the real operating system. I’ve started uploading this project to GitHub, where users can access and explore the emulator. It offers a unique way to experience the look and feel of Windows 10 on a web platform.